+++ WebUI/Models/RoomNumberDropItemView.cs
@@ -3,7 +3,7 @@ namespace WebUI.Models
 {
     public class RoomNumberDropItemView
     {
-        public int? Id { get; set; }
+        public byte? Id { get; set; }
         public string Name { get; set; }
     }
 }
